requirements-quality.txt
pytest>=9.0.0,<10.0.0
croniter
requests
python-dotenv
ddgs
psutil
